requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
kafka-python==2.0.2
protobuf==4.21.12
requests==2.28.1
urllib3==1.26.13
